Moves Direction enum
